+++ src/routes/main/dps/settings/settings.ts
@@ -1,10 +1,10 @@
-import type { ShortcutSettingId } from "$lib/settings-store";
+import type { ShortcutCommandSettingId } from "$lib/settings-store";
 
 export type BaseInputs = BaseInput[];
 
 /** Common base for all settings */
 export interface BaseInput {
-  id: ShortcutSettingId;
+  id: ShortcutCommandSettingId;
   label: string;
   description?: string;
 }