+++ src/routes/main/tool-sidebar.svelte
@@ -58,9 +58,9 @@
   }
 
   function formatAppVersion(version: string): string {
-    return version.replace(/-beta\.?(\d+)(rc)?$/i, (_, number, rc) =>
-      `_beta${number}${rc ? "RC" : ""}`,
-    );
+    return version
+      .replace(/-beta\.?(\d+)(rc)?$/i, (_, number, rc) => `_beta${number}${rc ? "RC" : ""}`)
+      .replace(/-rc\.?(\d+)$/i, (_, number) => `_RC${number}`);
   }
 
   async function toggleOverlayWindow() {